+++ internal/provider/repository/format/yum.go
@@ -394,9 +394,8 @@ func yumSchemaAttributes(isHosted bool) map[string]tfschema.Attribute {
 		"yum": schema.ResourceOptionalSingleNestedAttribute(
 			"YUM specific configuration for this Repository",
 			map[string]tfschema.Attribute{
-				"key_pair": schema.ResourceOptionalStringWithPlanModifier(
+				"key_pair": schema.ResourceOptionalString(
 					"PGP signing key pair (armored private key e.g. gpg --export-secret-key --armor)",
-					stringplanmodifier.UseStateForUnknown(),
 				),
 				"passphrase": schema.ResourceSensitiveOptionalStringWithPlanModifier(
 					"Passphrase to access PGP signing key",